Various fixes to LilyPond 2.23.x installation
Enable the Cairo backend.

Compile Scheme files into bytecode to offset the performance
penalty of Guile 2.2, see:
https://lilypond.org/doc/v2.23/Documentation/changes-big-page.html

Patch Guile so that Scheme files are never recompiled to .go files.
The problem is that Flatpak doesn't preserve timestamps:
https://github.com/flatpak/flatpak/issues/3064

Without this patch to Guile, LilyPond hangs forever while
trying to rebuild files which cannot be modified in the sandbox.